# app/static/docs/api.yaml
# this is the sensei API
swagger: '2.0'
info:
  title: Sensei API
  description: Augmenting classroom observation.
  version: "1.0.0"
# array of all schemes that your API supports
schemes:
  - https
# will be prefixed to all paths
basePath: /api/v1
produces:
  - application/json
securityDefinitions:
  basicAuth:
    type: basic
    description: HTTP Basic Authentication. Requires `HTTPS`
  apiKeyAuth:
    type: apiKey
    name: X-SenseiToken
    in: header
paths:
  /radio_observations:
    post:
      summary: Upload radio observations.
      description: |
        The Radio Observations endpoint provides a way to upload radio observations from sensors to the sensei platform.
      consumes:
        - application/json
      parameters:
        - name: events
          in: body
          description: List of radio observations.
          required: true
          schema:
            type: array
            items:
              $ref: '#/definitions/RadioObservation'
      tags:
        - RadioObservation
      security:
        - basicAuth: []
        - apiKeyAuth: []
      responses:
        201:
          description: Upload accepted
        401:
          description: Unauthorized
  /sensor_mappings:
    post:
      summary: Update mapping of sensor ids to children, teachers, materials, and classroom areas
      description: |
        This endpoint is used to create and update mappings of sensors to physical entities, such as children, teachers, materials, or areas. To associate a sensor with a new entity, start_time and end_time should not be specified. The server will end any existing mapping of this sensor. To mark a sensor as not associated with any entity, the entity type and entity id fields should be left blank.
      consumes:
        - application/json
      parameters:
        - name: mapping
          in: body
          required: true
          schema:
            type: array
            items:
              $ref: '#/definitions/SensorMapping'
      tags:
        - SensorMapping
      security:
        - basicAuth: []
      responses:
        201:
          description: Sensor mapping update accepted
        401:
          description: Unauthorized
    get:
      summary: Returns the current sensor mappings for the specified classroom.  Only active mappings will be returned.
      produces:
        - application/json
      parameters:
        - name: classroom_id
          in: query
          description: The id of the classroom to retrieve sensor mappings for.
          required: true
          type: integer
      tags:
        - SensorMapping
      security:
        - basicAuth: []
      responses:
        200:
          description: mappings
          schema:
            type: array
            items:
              $ref: '#/definitions/SensorMapping'
        401:
          description: Unauthorized
  /areas:
    post:
      summary: Create new area
      consumes:
        - application/json
      parameters:
        - name: area
          in: body
          required: true
          schema:
            $ref: '#/definitions/Area'
      tags:
        - Area
      security:
        - basicAuth: []
      responses:
        201:
          description: Area created
        401:
          description: Unauthorized
    get:
      summary: Returns the areas defined for the specified classroom.
      produces:
        - application/json
      parameters:
        - name: classroom_id
          in: query
          description: The id of the classroom to retrieve areas for.
          required: true
          type: integer
      tags:
        - Area
      security:
        - basicAuth: []
      responses:
        '200':
          description: Areas
          schema:
            type: array
            items:
              $ref: '#/definitions/Area'
        401:
          description: Unauthorized
  /areas/{id}:
    put:
      summary: Update an area's name.
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          description: The id of the area to update.
          required: true
          type: integer
        - name: mapping
          in: body
          description: Area.
          required: true
          schema:
            $ref: '#/definitions/Area'
      tags:
        - Area
      security:
        - basicAuth: []
      responses:
        '200':
          description: Area updated
          schema:
            $ref: '#/definitions/Area'
        401:
          description: Unauthorized
    delete:
      summary: Delete an area.
      parameters:
        - name: id
          in: path
          description: The id of the area to delete.
          required: true
          type: integer
      tags:
        - Area
      security:
        - basicAuth: []
      responses:
        '200':
          description: Area deleted
        401:
          description: Unauthorized
  /materials:
    post:
      summary: Create new material record
      consumes:
        - application/json
      parameters:
        - name: material
          in: body
          required: true
          schema:
            $ref: '#/definitions/Material'
      tags:
        - Material
      security:
        - basicAuth: []
      responses:
        201:
          description: Material created
        401:
          description: Unauthorized
    get:
      summary: Returns the materials defined for the specified classroom.
      produces:
        - application/json
      parameters:
        - name: classroom_id
          in: query
          description: The id of the classroom to retrieve materials for.
          required: true
          type: integer
      tags:
        - Material
      security:
        - basicAuth: []
      responses:
        '200':
          description: Materials
          schema:
            type: array
            items:
              $ref: '#/definitions/Material'
        401:
          description: Unauthorized
  /materials/{id}:
    put:
      summary: Update a material's name.
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          description: The id of the material to update.
          required: true
          type: integer
        - name: mapping
          in: body
          required: true
          schema:
            $ref: '#/definitions/Material'
      tags:
        - Material
      security:
        - basicAuth: []
      responses:
        '200':
          description: Material updated
          schema:
            $ref: '#/definitions/Material'
        401:
          description: Unauthorized
    delete:
      summary: Delete a material.
      parameters:
        - name: id
          in: path
          description: The id of the material to delete.
          required: true
          type: integer
      tags:
        - Material
      security:
        - basicAuth: []
      responses:
        '200':
          description: Material deleted
        401:
          description: Unauthorized
  /classrooms:
    get:
      summary: Returns the classrooms accessible to this user.
      produces:
        - application/json
      tags:
        - Classroom
      security:
        - basicAuth: []
      responses:
        '200':
          description: Classrooms
          schema:
            type: array
            items:
              $ref: '#/definitions/Classroom'
        401:
          description: Unauthorized

definitions:
  SensorMapping:
    type: object
    description: SensorMapping objects describe which sensors are mapped to which things, over a specific period of time.
    properties:
      classroom_id:
        type: integer
        description: Unique identifier for the classroom that the sensors involved belong to.
      sensor_id:
        type: integer
        description: Classroom specific sensor id of the sensor that is being mapped (0-255).
      start_time:
        type: string
        description: The starting time of the mapping, in iso8601 format. Example '2002-12-25T12:15:05'. If not set during a create, the current time will be used.
      end_time:
        type: string
        description: The ending time of the mapping, in iso8601 format. Example '2002-12-25T12:15:05'.  If not set, this indicates the current mapping.
      entity_type:
        type: string
        enum: [ "child", "teacher", "material", "area"]
        description: The type of mapping
      entity_id:
        type: integer
        description: The foreign key to
  RadioObservation:
    type: object
    description: Radio observations occur when when sensor receives a radio packet from another sensor. The receiver records the observed RSSI of the packet, along with the sender's id and a timestamp.
    properties:
      classroom_id:
        type: integer
        description: Unique identifier for the classroom that the sensors involved belong to.
      local_id:
        type: integer
        description: Classroom specific sensor id of the sensor that received and recorded the triggering packet (0-255).
      remote_id:
        type: integer
        description: The id of the sensor that sent the triggering packet (0-255).
      observed_at:
        type: string
        description: The timestamp of the event, in iso8601 format. Example '2002-12-25T12:15:05'
      rssi:
        type: number
        description: The signal strength, represented as a negative number of decibels.
  Area:
    type: object
    description: Describes an area of the classroom.
    properties:
      classroom_id:
        type: integer
        description: Unique identifier for the classroom that the area describes.
      name:
        type: string
        description: Name of the area.
  Material:
    type: object
    description: Classroom materials that children engage with for lessons.
    properties:
      classroom_id:
        type: integer
        description: Unique identifier for the classroom that the material is in.
      name:
        type: string
        description: Name of the material.
      lesson_id:
        type: integer
        description: Reference linking this material to a lesson in Transparent Classroom.
  Classroom:
    type: object
    description: Describes a classroom. This endpoint is currently backed by Transparent Classroom.
    properties:
      id:
        type: integer
        description: Unique identifier for the classroom.
      name:
        type: string
        description: Name of the classroom.
      lesson_set_id:
        type: integer
        description: Unique identifier for this class' lesson set.